==> R/05_analyses.Rmd <==
---
title: "Analyses"
author : "Jeremy Wicquart"
date: "`r Sys.Date()`"
output: 
  html_document:
    fig_caption: yes
    theme: "cosmo"
    highlight: tango
    toc: true
    toc_float:
      collapsed: false
      smooth_scroll: false
    toc_depth: 4
---

```{r setup, include=FALSE}

knitr::opts_chunk$set(echo = TRUE, message = FALSE, warning = FALSE, fig.align = "center")

```

# Defining functions and packages

```{r base}

# 1. Source functions ----

source("functions/graphical_par.R")
source("functions/theme_map.R")
source("functions/theme_graph.R")
source("functions/dhw_calculation.R")
source("functions/ts_decomp.R")
source("functions/lm_ts.R")
source("functions/mann_kendall_trend.R")

# 2. Required packages ----

library(tidyverse) # Core tidyverse packages
library(sf) # To plot maps
sf_use_s2(FALSE) # Switch from S2 to GEOS
library(lubridate)
library(ggpubr)
library(magick) # To create GIF
library(trend) # For Mann-Kendall test

# 3. Set theme_graph() as the default ggplot theme ----

theme_set(theme_map())

# 4. Define the CRS ----

crs_selected <- "+proj=eqearth"

```

# 1. Misc. calculation on SST

```{r}

# 1. Load data ----

load("./../data/02_sst-extracted.RData")

# 2. DHW and SST anomaly calculation ----

results_sst <- dhw_calculation(results_sst)

# 3. Decompose the time-series into trend, seasonality and random ----

results_sst <- results_sst %>% 
  group_by(ecoregion) %>% 
  do(ts_decomp(.))

# 4. Fit linear model ----

results_sst <- results_sst %>% 
  group_by(ecoregion) %>% 
  do(lm_ts(.))

```

# 2. SST description plot by ecoregion

# 2.1 SST, SST anomaly and DHW

```{r}

# 1. SST trends ----

# A. Create the empty list --

plot_sst <- list() # Empty list

# B. Loop --

for (i in unique(results_sst$ecoregion)) {
  
  data_i <- results_sst %>% 
    filter(ecoregion == i)
  
  plot_i <- ggplot(data = data_i) +
    geom_path(aes(x = date, y = sst)) +
    labs(x = "Year", y = "SST (°C)", title = unique(data_i$ecoregion)) +
    geom_hline(yintercept = unique(data_i$threshold), linetype = "dashed", color = "red") +
    geom_path(aes(x = date, y = sst_fit), color = "red", size = 1) +
    theme_graph() +
    theme(axis.title.x = element_blank()) +
    lims(x = c(as.Date("1985-01-01"), as.Date("2020-12-31")))

  plot_sst[[i]] <- plot_i # add each plot into plot list
  
}

# 2. SST anomaly trends ----

# A. Create the empty list --

plot_anom <- list() # Empty list

# B. Loop --

for (i in unique(results_sst$ecoregion)) {
  
  data_i <- results_sst %>% 
    filter(ecoregion == i)
  
  plot_i <- ggplot(data = data_i, aes(x = date, y = sst_anomaly_mean)) +
    geom_path() +
    geom_ribbon(aes(ymin = 0, ymax = sst_anomaly_mean, fill = sst_anomaly_type), alpha = 0.5) +
    scale_fill_identity() +
    labs(x = "Year", y = "SST anomaly (°C)") +
    theme_graph() +
    theme(axis.title.x = element_blank()) +
    lims(x = c(as.Date("1985-01-01"), as.Date("2020-12-31")))

  plot_anom[[i]] <- plot_i # add each plot into plot list
  
}

# 3. DHW trends ----

# A. Create the empty list --

plot_dhw <- list() # Empty list

# B. Loop --

for (i in unique(results_sst$ecoregion)) {
  
  data_i <- results_sst %>% 
    filter(ecoregion == i) %>% 
    drop_na(dhw)
  
  plot_i <- ggplot(data = data_i, aes(x = date, y = dhw)) +
    geom_path() +
    geom_ribbon(aes(ymin = 0, ymax = dhw), fill = "#ec644b", alpha = 0.5) +
    labs(x = "Year", y = "DHW (°C-weeks)") +
    theme_graph() +
    lims(x = c(as.Date("1985-01-01"), as.Date("2020-12-31")))

  plot_dhw[[i]] <- plot_i # add each plot into plot list
  
}

# 4. Loop to combine and export plots ----

for (i in names(plot_sst)) {
  
  plot_i <- ggarrange(plotlist = c(plot_sst[i], plot_anom[i], plot_dhw[i]),
                      ncol = 1, nrow = 3, heights = c(1.1, 1, 1), align = "v")
  
  ggsave(plot = plot_i, 
         filename = paste0("../figs/01_sst-anom-dhw_", str_replace_all(tolower(i), " ", "-"), ".png"),
         width = 6, height = 11)
  
}

# 5. Remove useless objects ----

rm(plot_sst, plot_anom, plot_dhw, plot_i)

```

# 2.2 Time-series decomposition

```{r}

# 1. Make a loop to produce each figure by ecoregion ----

for (i in unique(results_sst$ecoregion)) {
  
  palette_facet <- data.frame(type = c("observed", "trend", "seasonal", "random"), 
                              color = c("#d64541", "#2abb9b", "#3498db", "#be90d4"))
  
  data_i <- results_sst %>% 
    filter(ecoregion == i) %>% 
    pivot_longer(., cols = c(sst, trend, seasonal, random), values_to = "value", names_to = "type") %>% 
    mutate(type = str_replace_all(type, c("sst" = "observed"))) %>% 
    left_join(., palette_facet) %>% 
    mutate(type = factor(type, levels = c("observed", "trend", "seasonal", "random")))

  plot_i <- ggplot(data_i, aes(x = date, y = value, color = color)) +
    geom_path(show.legend = FALSE) +
    facet_wrap(~type, scales = "free", ncol = 1) +
    labs(x = "Year", y = "SST (°C)", title = as.character(i)) +
    theme_graph() +
    scale_color_identity()
  
  ggsave(filename = paste0("./../figs/04_sst-time-series-decomposed-", 
                           str_replace_all(tolower(i), " ", "-"),
                           ".png"),
         width = 6, height = 11)

}

# 2. Remove useless objects ----

rm(data_i, i)

```

# 3. Table of SST parameters

```{r}

# 1. Calculate significance of the SST slope (Mann-Kendall) ----

results_sst_mk <- results_sst %>% 
  group_by(ecoregion) %>% 
  do(mann_kendall_trend(.))

# 2. Calculate SST difference between 2000 and 1985 ----  

results_sst_table <- results_sst %>% 
  select(ecoregion, mean, date, threshold, sst_fit) %>% 
  filter(date == max(date) | date == min(date)) %>% 
  mutate(date = str_replace_all(date, c("1985-01-01" = "before",
                                        "2020-12-31" = "after"))) %>% 
  pivot_wider(values_from = sst_fit, names_from = date) %>% 
  mutate(diff_sst = after-before) %>% 
  select(-before, -after) %>% 
  left_join(., results_sst_mk)

# 3. Remove useless objects ----

rm(results_sst_mk)

```

```{r}

# 1. Make the figure ----

ggplot(data = results_sst_table) +
  geom_vline(xintercept = 0, colour = "#2e3131", linetype = "dotted") +
  geom_histogram(aes(x = diff_sst), bins = 15, fill = "#dadfe1", color = "#2e3131") +
  geom_jitter(aes(x = diff_sst, y = -4, fill = diff_sst), shape = 21, 
              show.legend = FALSE, size = 3, height = 1) +
  geom_boxplot(aes(x = diff_sst, y = -4), fill = "#dadfe1", color = "#2e3131", outlier.size = 0, 
               outlier.color = NA, width = 4, alpha = 0.3) +
  scale_fill_gradientn(colours = c("#22a7f0", "white", "#fad859", "#e47833", "#f03434"),
                       limits = c(-0.5, 1.5)) +
  annotate("segment", 
           x = mean(results_sst_table$diff_sst),
           xend = mean(results_sst_table$diff_sst), y = 45, yend = 40, colour = "black",
           arrow = arrow(length = unit(0.05, "inches")), lineend = "round", linejoin = "round") +
  annotate("text", x = mean(results_sst_table$diff_sst), y = 47, 
           label = paste0("Mean = ", round(mean(results_sst_table$diff_sst), 3),"°C"), size = 3.5, family = font_choose_graph) +
  theme_graph() +
  lims(x = c(-0.5, 1.5), y = c(NA, 50)) +
  labs(x = expression(Delta*" SST 2020-1985"), y = "Number of ecoregions")

# 2. Save the figure ----

ggsave("./../figs/04_distribution-sst-diff-2020-1985.png", width = 7, height = 4, dpi = 600)

```




































```{r}

ggplot(data = results_sst, aes(x = date, y = sst_anomaly_mean, group = ecoregion)) +
    annotate("rect", xmin = as.Date("1997-01-01"), xmax = as.Date("1999-12-31"), 
             ymin = 0, ymax = 1, fill = "#ec644b", alpha = 0.5) +
    annotate("rect", xmin = as.Date("2009-01-01"), xmax = as.Date("2011-12-31"), 
             ymin = 0, ymax = 1.1, fill = "#ec644b", alpha = 0.5) +
    annotate("rect", xmin = as.Date("2015-01-01"), xmax = as.Date("2017-12-31"), 
             ymin = 0, ymax = 1.35, fill = "#ec644b", alpha = 0.5) +
    geom_path() +
    labs(x = "Year", y = "SST anomaly (°C)") +
    theme_graph() +
    lims(x = c(as.Date("1985-01-01"), as.Date("2020-12-31")))

```

```{r}

ggplot(data = results_sst, aes(x = date, y = dhw, group = ecoregion)) +
    geom_path() +
    labs(x = "Year", y = "DHW (°C-weeks)") +
    theme_graph() +
    lims(x = c(as.Date("1985-01-01"), as.Date("2020-12-31")))

```

# 2. Dataviz

# 2.0 Background maps

```{r}

# 1. Load background maps ----

data_map <- read_sf("./../data/00_natural-earth-data/ne_10m_land/ne_10m_land.shp") %>% 
  st_transform(crs = 4326) %>% 
  st_transform(crs = crs_selected)

# 2. Create the border of background map ----

lats <- c(90:-90, -90:90, 90)
longs <- c(rep(c(180, -180), each = 181), 180)

background_map_border <- list(cbind(longs, lats)) %>%
  st_polygon() %>%
  st_sfc(crs = 4326) %>% 
  st_sf() %>%
  st_transform(crs = crs_selected)

```

# 2.1 Coral reefs distribution

```{r}

# 1. Load Reefs at Risk data ----

data_rar <- read_sf("./../data/02_reefs-at-risk_reef-data/reef_500_poly.shp") %>% 
  st_transform(crs = 4326) %>% 
  st_wrap_dateline(options = c("WRAPDATELINE=YES")) %>% 
  st_transform(crs = crs_selected)

# 2. Make the figure ----

ggplot() +
  # Background map border
  geom_sf(data = background_map_border, fill = "#56B4E950", color = "grey30", size = 0.5/.pt) +
  # Coral distribution
  geom_sf(data = data_rar, col = "#2abb9b", size = 2) +
  # Background map
  geom_sf(data = data_map, fill = col_fill_map, col = col_color_map)

# 3. Save the figure ----

ggsave("./../figs/01_coral-reefs-distribution.png", width = 7, height = 4, dpi = 600)

```

# 2.2 Coral reefs ecoregions

```{r}

# 1. Load ecoregions data ----

load("./../data/ecoregions.RData")
load("./../data/data_joined.RData")

ecoregions <- ecoregions %>% 
  filter(Ecoregion %in% unique(data_joined$Ecoregion)) %>% 
  st_transform(crs = 4326) %>% 
  st_wrap_dateline(options = c("WRAPDATELINE=YES")) %>% 
  st_make_valid() %>% 
  st_wrap_dateline(options = c("WRAPDATELINE=YES")) %>% 
  st_make_valid() %>% 
  st_transform(crs = crs_selected) %>% 
  st_wrap_dateline(options = c("WRAPDATELINE=YES")) %>% 
  st_make_valid()

# 2. Make the figure ----

ggplot() +
  # Background map border
  geom_sf(data = background_map_border, fill = "#56B4E950", color = "grey30", size = 0.5/.pt) +
  # Coral reefs ecoregions
  geom_sf(data = ecoregions, aes(fill = Ecoregion), show.legend = FALSE) +
  # Background map
  geom_sf(data = data_map, fill = col_fill_map, col = col_color_map)

# 3. Save the figure ----

ggsave("./../figs/02_coral-reefs-ecoregions.png", width = 7, height = 4, dpi = 600)

```

# 2.3 SST increase 2020-1985

```{r}

# 1. Join with ecoregion polygon ----

ecoregions_delta_sst <- ecoregions %>% 
  select(Ecoregion) %>% 
  left_join(results_sst_table, ., by = c("ecoregion" = "Ecoregion")) %>% 
  st_as_sf()

# 2. Make the figure ----

ggplot() +
  # Background map border
  geom_sf(data = background_map_border, fill = "#56B4E950", color = "grey30", size = 0.5/.pt) +
  # Coral reefs ecoregions
  geom_sf(data = ecoregions_delta_sst, aes(fill = diff_sst), color = NA) +
  # Background map
  geom_sf(data = data_map, fill = col_fill_map, col = col_color_map) +
  scale_fill_gradientn(colours = c("#22a7f0", "white", "#fad859", "#e47833", "#f03434"),
                       limits = c(-0.5, 1.5)) +
  guides(fill = guide_colourbar(title = expression(Delta*" SST 2020-1985"),
                                title.position = "top",
                                barwidth = 15,
                                barheight = 1,
                                frame.colour = "black",
                                frame.linewidth = 1,
                                ticks.colour = "black"))

# 3. Save the figure ----

ggsave("./../figs/03_map-delta-sst-2020-1985.png", width = 7, height = 4, dpi = 600)

```




































# 2.3 Max DHW by ecoregion and year

```{r}

# 1. Calculate max DHW by ecoregion and year ----

results_sst <- results_sst %>% 
  mutate(year = year(date)) %>% 
  group_by(year, ecoregion) %>% 
  summarise(max_dhw = max(dhw, na.rm = TRUE))

# 2. Join with ecoregion polygon ----

ecoregions2 <- ecoregions %>% 
  select(Ecoregion) %>% 
  left_join(results_sst, ., by = c("ecoregion" = "Ecoregion")) %>% 
  st_as_sf()

# 3. Make the figure ----

ggplot() +
  # Background map border
  geom_sf(data = background_map_border, fill = "#56B4E950", color = "grey30", size = 0.5/.pt) +
  # Coral reefs ecoregions
  geom_sf(data = ecoregions2, aes(fill = max_dhw), color = NA) +
  facet_wrap(~year, ncol = 4) +
  # Background map
  geom_sf(data = data_map, fill = col_fill_map, col = col_color_map) +
  scale_fill_gradientn(colours = c("white", "#03a678", "#446cb3", "#e47833", "#d91e18")) +
  guides(fill = guide_colourbar(title = "Maximum DHW (°C-weeks)",
                                title.position = "top",
                                barwidth = 15,
                                barheight = 1,
                                frame.colour = "black",
                                frame.linewidth = 1,
                                ticks.colour = "black"))

# 4. Save the figure ----

ggsave("./../figs/03_max-dhw-by-ecoregion-and-year.png", width = 22, height = 25, dpi = 600)

```

# GIF

```{r}

# 1. Loop to produce each individual figure by year ----

for(i in unique(ecoregions2$year)){
  
  data_i <- ecoregions2 %>% 
    filter(year == i)
  
  plot_i <- ggplot() +
  # Background map border
  geom_sf(data = background_map_border, fill = "#56B4E950", color = "grey30", size = 0.5/.pt) +
  # Coral reefs ecoregions
  geom_sf(data = data_i, aes(fill = max_dhw), color = NA) +
  # Background map
  geom_sf(data = data_map, fill = col_fill_map, col = col_color_map) +
  labs(title = as.character(i)) +
  theme(legend.position = "bottom",
        plot.title = element_text(hjust = 0.5),
        panel.background = element_rect(fill = "white", colour = NA),
        plot.background = element_rect(fill = "white")) +
  scale_fill_gradientn(colours = c("white", "#03a678", "#446cb3", "#e47833", "#d91e18"), limits = c(0, 8)) +
  guides(fill = guide_colourbar(title = "Maximum DHW (°C-weeks)",
                                title.position = "top",
                                barwidth = 15,
                                barheight = 1,
                                frame.colour = "black",
                                frame.linewidth = 1,
                                ticks.colour = "black"))
  
  ggsave(plot = plot_i, 
         filename = paste0("./../figs/03_max-dhw-by-ecoregion-gif", as.character(i),".png"), 
         width = 8, height = 5, dpi = 600)

}

# 2. List of individual figures ----

imgs <- list.files("./../figs/", pattern = "max-dhw-by-ecoregion-gif", full.names = TRUE)
img_list <- lapply(imgs, image_read)

# 3. Join the images together ----

img_joined <- image_join(img_list)

# 4. Create the animation at 2 FPS ----

img_animated <- image_animate(img_joined, fps = 1)

# 5. Export the GIF ----

image_write(image = img_animated, path = "./../figs/03_gif-max-dhw-by-ecoregion.gif")

```




























# 2.3 SST

```{r}




```

# 2.4 TS

```{r}

# 1. Load TS data ----

load(file = "./../data/03_tropical-storms_raw/data_cyclones.RData")

data_cyclones <- data_cyclones %>% 
  group_by(ts_id) %>% 
  mutate(wind_speed = max(wind_speed)) %>% 
  st_transform(crs = 4326) %>% 
  st_make_valid() %>% 
  # Transform points to lines
  group_by(ts_id, wind_speed) %>%
  summarise(do_union = FALSE) %>%
  st_cast("LINESTRING") %>% 
  # Remove linestring with only one point (else error with the st_intersection() function)
  mutate(n = str_count(geometry, ",")) %>% 
  filter(n > 1) %>% 
  select(-n) %>% 
  st_wrap_dateline(options = c("WRAPDATELINE=YES")) %>% 
  st_make_valid() %>% 
  st_wrap_dateline(options = c("WRAPDATELINE=YES")) %>% 
  st_make_valid() %>% 
  st_transform(crs = crs_selected) %>% 
  st_wrap_dateline(options = c("WRAPDATELINE=YES")) %>% 
  st_make_valid()

# 2. Make the figure ----

ggplot() +
  # Background map border
  geom_sf(data = background_map_border, fill = "#56B4E950", color = "grey30", size = 0.5/.pt) +
  # Tropical storms data
  geom_sf(data = data_cyclones, aes(col = wind_speed)) +
  # Background map
  geom_sf(data = data_map, fill = col_fill_map, col = col_color_map) +
  scale_colour_gradientn(colours = c("#03a678", "#446cb3", "#e47833", "#d91e18"), limits = c(0, 350)) +
  guides(col = guide_colourbar(title = expression(paste("Maximum wind speed (km.", h^{-2}, ")")),
                               title.position = "top",
                               barwidth = 15,
                               barheight = 1,
                               frame.colour = "black",
                               frame.linewidth = 1,
                               ticks.colour = "black"))

# 3. Save the figure ----

ggsave("./../figs/03_tropical-storms-paths.png", width = 7, height = 5, dpi = 600)

```


```{r}

load(file = "./../data/03_tropical-storms_raw/data_cyclones.RData")

data_cyclones <- data_cyclones %>% 
  group_by(ts_id, saffir) %>% 
  mutate(wind_speed = max(wind_speed),
         saffir = max(saffir)) %>% 
  ungroup() %>% 
  drop_na(saffir) %>%
  mutate(year = year(time)) %>% 
  group_by(year, saffir) %>% 
  summarise(n_event = n()) %>% 
  ungroup() %>% 
  tidyr::complete(year, saffir, fill = list(n_event = 0)) %>% 
  filter(saffir != 0) %>% 
  mutate(saffir = as.factor(saffir)) %>% 
  filter(year >= 1960)


# test 

library(ggstream)

ggplot(data = data_cyclones, aes(x = year, y = n_event, fill = saffir)) +
  geom_bar(stat = "identity") +
  theme_graph()

ggplot(data = data_cyclones, aes(x = year, y = n_event, fill = saffir)) +
  geom_stream(color = "black", lwd = 0.1) +
  theme_graph()



```

```{r fig.width=15}

library(ggridges)

load(file = "./../data/03_tropical-storms_raw/data_cyclones.RData")

data_cyclones <- data_cyclones %>% 
  group_by(ts_id) %>% 
  mutate(wind_speed = max(wind_speed),
         year = year(time)) %>% 
  group_by(ts_id, name, year, wind_speed) %>% 
  distinct() %>% 
  filter(year >= 1970)

ggplot(data = data_cyclones, aes(x = wind_speed, y = as.factor(year))) +
  geom_density_ridges(fill = "#3498db") +
  theme_graph()


```

```{r}

load(file = "./../data/03_tropical-storms_raw/data_cyclones.RData")

data_cyclones <- data_cyclones %>% 
  group_by(ts_id) %>% 
  mutate(wind_speed = max(wind_speed),
         year = year(time)) %>% 
  group_by(ts_id, name, year, wind_speed) %>% 
  distinct() %>% 
  filter(year >= 1970)

ggplot(data = data_cyclones, aes(x = year, y = wind_speed, group = year)) +
  geom_violin(fill = "#3498db", alpha = 0.25, col = NA) +
  geom_point(shape = 1, col = "#5c97bf", alpha = 0.5, size = 1) +
  theme_graph() +
  labs(x = "Year", 
       y = expression(paste("Maximum wind speed (km.", h^{-2}, ")")))

```

# Composed maps

```{r}

# 1. Load ecoregions data ----

load("./../data/ecoregions.RData")
load("./../data/data_joined.RData")

ecoregions <- ecoregions %>% 
  filter(Ecoregion %in% unique(data_joined$Ecoregion)) %>% 
  st_transform(crs = 4326) %>% 
  st_wrap_dateline(options = c("WRAPDATELINE=YES")) %>% 
  st_make_valid() %>% 
  st_wrap_dateline(options = c("WRAPDATELINE=YES")) %>% 
  st_make_valid()

ecoregions2 <- ecoregions %>% 
  mutate(region = case_when(Ecoregion %in% c("Bermuda",
                                             "Belize and west Caribbean",
                                             "Cuba and Cayman Islands",
                                             "Bahamas and Florida Keys",
                                             "Bay of Campeche, Yucatan, Gulf of Mexico",
                                             "Flower Garden Banks, Gulf of Mexico",
                                             "Hispaniola, Puerto Rico and Lesser Antilles",
                                             "Netherlands Antilles and south Caribbean",
                                             "Jamaica") ~ "Caribbean")) %>% 
  filter(region == "Caribbean")


# 1. Load background maps ----

data_map <- read_sf("./../data/00_natural-earth-data/ne_10m_land/ne_10m_land.shp") %>% 
  st_transform(crs = 4326)

# 2. Create the border of background map ----

lats <- c(90:-90, -90:90, 90)
longs <- c(rep(c(180, -180), each = 181), 180)

background_map_border <- list(cbind(longs, lats)) %>%
  st_polygon() %>%
  st_sfc(crs = 4326) %>% 
  st_sf()




# 2. Make the figure ----

ggplot() +
  # Background map border
  geom_sf(data = background_map_border, fill = "#56B4E950", color = "grey30", size = 0.5/.pt) +
  # Coral reefs ecoregions
  geom_sf(data = ecoregions2, aes(fill = Ecoregion)) +
  # Background map
  geom_sf(data = data_map, fill = col_fill_map, col = col_color_map) +
  coord_sf(xlim = c(-100, -40), ylim = c(7, 35)) +
  theme(legend.position = "right",
        legend.direction = "vertical",
        legend.text = element_text(size = 8),
        legend.justification = "left")

# 3. Save the figure ----

ggsave("./../figs/02_coral-reefs-ecoregions.png", width = 7, height = 4, dpi = 600)


```

# Reproducibility

```{r reprod}

# 1. Reproducibility ----

sessionInfo()

```

---
Jeremy WICQUART | jeremywicquart@gmail.com | `r format(Sys.time())`
